Move Clock to common so other artifacts can override or inject
