+++ sdk/src/main/java/io/opentelemetry/sdk/internal/TestClock.java
@@ -16,6 +16,7 @@
 
 package io.opentelemetry.sdk.internal;
 
+import io.opentelemetry.sdk.common.Clock;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.concurrent.GuardedBy;
 import javax.annotation.concurrent.ThreadSafe;
